# specs/001-student-management/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Student Management System API
  version: 1.0.0
  description: REST API for student enrollment, viewing, and record management
  contact:
    name: Student Management System
    email: admin@schoolsystem.com
servers:
  - url: http://localhost:8000
    description: Development server
paths:
...
  /api/v1/students:
    get:
      summary: List all students
      operationId: listStudents
      tags:
        - students
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
            description: Page number for pagination
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            default: 20
            description: Number of items per page
        - name: search
          in: query
          required: false
          schema:
            type: string
            description: Search students by first or last name (satisfies FR-003)
        - name: gradeLevel
          in: query
          required: false
          schema:
            type: string
            maxLength: 20
            description: Filter by grade level label (free text, e.g. "9" or "K-1")
        - name: active
          in: query
          required: false
          schema:
            type: boolean
            description: Filter by active status
      responses:
        '200':
          description: Successful response with student list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    description: >-
                      Student rows. Admins receive email/phone/date_of_birth;
                      teachers receive the same shape with personal contact info
                      nulled (student email/phone are admin-only).
                    items:
                      $ref: '#/components/schemas/StudentDetail'
                  meta:
                    type: object
                    properties:
                      page:
                        type: integer
                      pageSize:
                        type: integer
                      total:
                        type: integer
        '401':
          description: Unauthorized - invalid or missing authentication token
        '403':
          description: Forbidden - insufficient permissions

    post:
      summary: Enroll new student
      operationId: enrollStudent
      tags:
        - students
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentEnrollment'
      responses:
        '201':
          description: Student successfully enrolled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentResponse'
        '400':
          description: Bad request - validation failed (e.g., duplicate, missing required fields)
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /api/v1/students/{student_id}:
    get:
      summary: Get student by ID
      operationId: getStudent
      tags:
        - students
      security:
        - bearerAuth: []
      description: >-
        Admins see the full profile. Teachers see the profile of students they
        teach only (else 403), with email/phone nulled — student contact info is
        admin-only. Parents get 403.
      parameters:
        - name: student_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response with student details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentDetail'
        '404':
          description: Student not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

    put:
      summary: Update student information
      operationId: updateStudent
      tags:
        - students
      security:
        - bearerAuth: []
      parameters:
        - name: student_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentUpdate'
      responses:
        '200':
          description: Student successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentResponse'
        '400':
          description: Bad request - validation failed
        '404':
          description: Student not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /api/v1/enrollments:
    post:
      summary: Create new enrollment record
      operationId: createEnrollment
      tags:
        - enrollments
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnrollmentCreate'
      responses:
        '201':
          description: Enrollment successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentResponse'
        '400':
          description: Bad request - validation failed (e.g., student already enrolled, invalid student_id)
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  /api/v1/courses/{course_id}/students:
    get:
      summary: List students enrolled in a course
      operationId: listCourseStudents
      tags:
        - courses
      security:
        - bearerAuth: []
      parameters:
        - name: course_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response with enrolled students
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/StudentSummary'
        '404':
          description: Course not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    put:
      summary: Replace the roster of students assigned to a course
      operationId: updateCourseStudents
      tags:
        - courses
      security:
        - bearerAuth: []
      parameters:
        - name: course_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseRosterUpdate'
      responses:
        '200':
          description: Course roster updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDetail'
        '400':
          description: Bad request - invalid or missing student IDs
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  /api/v1/students/{student_id}/parents:
    get:
      summary: List all parents linked to a student
      operationId: listStudentParents
      tags:
        - students
      security:
        - bearerAuth: []
      parameters:
        - name: student_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      description: >-
        Admins can look up any student's parents. Teachers can only look up the
        parents of students they teach (so they can reach a guardian about
        absences or poor performance); otherwise 403. Parents have their own
        portal and cannot view other people's records (403).
      responses:
        '200':
          description: Successful response with linked parents
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ParentSummary'
        '404':
          description: Student not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  /api/v1/parents/{parent_id}/students:
    get:
      summary: List students linked to a parent
      operationId: listParentStudents
      tags:
        - parents
      security:
        - bearerAuth: []
      parameters:
        - name: parent_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response with linked students
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/StudentSummary'
        '404':
          description: Parent not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    put:
      summary: Replace the students linked to a parent (admin-only)
      operationId: updateParentStudents
      tags:
        - parents
      security:
        - bearerAuth: []
      parameters:
        - name: parent_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParentStudentsUpdate'
      responses:
        '200':
          description: Parent's students updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/StudentSummary'
        '400':
          description: One or more students not found
        '404':
          description: Parent not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    StudentSummary:
      type: object
      properties:
        student_id:
          type: string
          format: uuid
        first_name:
          type: string
        last_name:
          type: string
        grade_level:
          type: string
        active:
          type: boolean
        enrollment_date:
          type: string
          format: date
    StudentDetail:
      type: object
      allOf:
        - $ref: '#/components/schemas/StudentSummary'
        - type: object
          properties:
            date_of_birth:
              type: string
              format: date
            email:
              type: string
              format: email
            phone:
              type: string
    StudentEnrollment:
      type: object
      required:
        - first_name
        - last_name
        - date_of_birth
        - grade_level
      properties:
        first_name:
          type: string
          maxLength: 100
        last_name:
          type: string
          maxLength: 100
        date_of_birth:
          type: string
          format: date
        grade_level:
          type: string
          minLength: 1
          maxLength: 20
          description: School-defined grade label (e.g. "9", "K-1", "O-Level")
        email:
          type: string
          format: email
          maxLength: 200
        phone:
          type: string
          maxLength: 50
    StudentUpdate:
      type: object
      properties:
        first_name:
          type: string
          maxLength: 100
        last_name:
          type: string
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 200
        phone:
          type: string
          maxLength: 50
    StudentResponse:
      type: object
      properties:
        success:
          type: boolean
        student:
          $ref: '#/components/schemas/StudentDetail'
        message:
          type: string
    EnrollmentCreate:
      type: object
      required:
        - student_id
        - enrolled_by
      properties:
        student_id:
          type: string
          format: uuid
        enrolled_by:
          type: string
          maxLength: 100
        status:
          type: string
          enum: [active, withdrawn, transferring]
          default: active
    EnrollmentResponse:
      type: object
      properties:
        success:
          type: boolean
        enrollment:
          $ref: '#/components/schemas/EnrollmentDetail'
        message:
          type: string
    EnrollmentDetail:
      type: object
      properties:
        enrollment_id:
          type: string
          format: uuid
        student_id:
          type: string
          format: uuid
        enrolled_by:
          type: string
        enrollment_date:
          type: string
          format: date
        status:
          type: string
...
    CourseSummary:
      type: object
      properties:
        course_id:
          type: string
          format: uuid
        name:
          type: string
        teacher_id:
          type: string
          format: uuid
        grade_level:
          type: string
        semester:
          type: string
        status:
          type: string
    CourseDetail:
      type: object
      allOf:
        - $ref: '#/components/schemas/CourseSummary'
        - type: object
          properties:
            max_students:
              type: integer
            students:
              type: array
              items:
                $ref: '#/components/schemas/StudentSummary'
...
    CourseRosterUpdate:
      type: object
      required:
        - student_ids
      properties:
        student_ids:
          type: array
          items:
            type: string
            format: uuid
    ParentSummary:
      type: object
      properties:
        parent_id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: boolean
...
    ParentStudentsUpdate:
      type: object
      properties:
        student_ids:
          type: array
          items:
            type: string
            format: uuid
          description: Replacement list of students linked to this parent
...